Murua_Con/app/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv 
pydantic_settings
service
pdf2image
fitz
pytesseract
doctr
docx2txt